CloudSimulator now does nothing if the cloud file is already initialized. CloudManipulator will still delete it if it is found.

+++ src/main/java/address/sync/cloud/CloudFileHandler.java
@@ -61,10 +61,11 @@ public class CloudFileHandler {
         }
     }
 
-    public void initializeCloudAddressBookFile(String addressBookName) throws IOException, DataConversionException,
+    public void initializeCloudAddressBookFile(String addressBookName, boolean shouldReplaceExisting) throws IOException, DataConversionException,
             IllegalArgumentException {
         File cloudFile = getCloudDataFile(addressBookName);
         if (cloudFile.exists()) {
+            if (!shouldReplaceExisting) return;
             cloudFile.delete();
         }
